requirements.txt
openai
python-dotenv
jsonschema
asteval
tiktoken